Witcher 3: Downscale WBuffer to improve SLI performance
Refer to the following issue:
https://github.com/bo3b/3Dmigoto/issues/53

stereo2mono requires a large amount of data to be synchronised between
GPUs in SLI, which at high resolutions (typically above 1920x1080) can
become a major bottleneck causing the framerate to take an enormous hit,
in some cases even making it unplayable.

To avoid this, downscale the WBuffer to 1/4 resolution prior to
using stereo2mono. We don't need a full resolution version of the
WBuffer to calculate the crosshair depth, so this is unlikely to have
any significant effect on the accuracy.