Add parsing routines for new [Profile] section
This adds parsing support for a new [Profile] section containing the
driver profile settings we want to apply.

The preferred way to use this new section is like so:

    [Profile]
    StereoProfile = 1
    StereoFlagsDX10 = 0x00004000
    StereoConvergence = 0.5
    Comments = "Far Cry Primal community fix. Please disable motion blur."
    Disable2DD = 1
    2DD_Notes = "Compatibility mode is enabled. To use the fix, please disable it with Ctrl+Alt+F11"

But for convenience it also has the flexibility to handle a profile
copied from the output of Geforce Profile Manager or dumped with
3DMigoto's new profile logging feature. This includes handling raw
Setting[String] ID_0x........ type names, and deciphering internal DWORD
values. Internal string values in this form are intentionally ignored
(and will sound the error beeps) due to the strong possibility they may
have become corrupt, and that if anyone actually wanted to override them
there is no way they would pass them with the internal setting flag.

Most of the parsing guesses the type based on the text, but since
integers and floating point values can be ambiguous, we also tag
settings that are known to be floating point values to ensure they will
always be correctly decoded.