Compare the driver profile to determine if it needs to be updated
This will compare the profile in the d3dx.ini with the one currently
loaded to the driver to decide whether it will need to update it or not.

It will decide to update it if the profile is missing or any settings
are missing or do not match, with one exception: Certain settings are
allowed to differ provided that they are user settings and not
predefined values. This allows things like the convergence to be
customised without insisting that we update the profile on every launch.
These settings will still be updated if we decide we need to update the
profile for any other reason, which allows us to force them to be
updated if we deem it necessary, by modifying an unrelated setting such
as by adding a version number to the Comments section.