Add code to update driver profile
If the profile manager determines that the profile needs to be updated,
it will now attempt to do so, creating the profile if necessary (using
the executable name with -3DMigoto appended as the profile name).

Once completed, it will log the profile again so that we can confirm if
the changes were committed as wanted, and sounding a new error tone
(Brnk, du-du-dunk) if it did not.