Copy 3DMigoto to a temp location if DLL's shortpath contains a space
In order to use rundll as our helper program to raise our privileges, we
cannot have a space or comma in the filename, otherwise rundll will
silently fail. The documentation recommends using the shortname, but
that only works if the file has a shortname, which is not always the
case. Eventually we might want to ship/embed our own helper utility to
take the place of rundll, but for now just copy the 3DMigoto DLL to the
temporary directory, which should have a much higher chance of not
having a space and a much higher chance of having a shortname.