Install the profile in admin helper mode
Now that the privilege elevation code is working, make the DLL actually
install the profile when called from rundll as a helper.

This follows a cut down version of the regular init procedure, but with
anything not relevant to the profiles removed. It also has some special
handling to use the executable path passed in on the command line to
find the game, rather than the executable it is loaded into (since that
is rundll). It uses an alternate log file d3d11_profile_log.txt since
the regular 3DMigoto will still be running waiting for the helper to
complete.