Move G to static memory
Globals is currently allocated on the heap, which is pointless given
that it is global, and fragile given that we now have a second entry
point for the profile helper that does not use the same init paths as
the regular dll. Change it to be statically allocated, but to avoid
needing to change every reference, keep the old "G" name as a pointer to
it.