Improve log messages for checking vs. updating profile
The check & update routines share common code, which was logging for
both and the messages were a tad confusing since they seemed to imply
errors in the update when in fact there were none.

Pass a flag to only log while checking if a change is needed, and add
separate logging to the update routine to indicate that the updates are
going through successfully.