Witcher 3: Decrypt profile and add setting name comments
After applying the Witcher 3 profile, the decrypted version was copied
out of the d3d11_log.txt and put back in the d3dx.ini. This decrypts the
internal settings in the profile and adds setting names to all
recognised settings.

I have gone out of my way to keep the order intact and removed the
UserSpecified flags for a minimal diff (use --word-diff for best
results), but this is not required to decrypt a profile in this manner.

I removed one of the duplicate 2DDConvergence settings (removed the
predefined, leaving the user value).