Fix off by one errors in max_executions/copies_per_frame
I happened to notice in frame analysis that my custom shader limited
with max_executions_per_frame=1 was not working as expected, and was
running twice instead. This is due to setting the current copy/execution
count to 0 on the first copy/execution, while it should be set to 1 to
count that copy/execution.

Also, add logging for when max_copies_per_frame is exceeded in the same
way exceeding max_executions_per_frame is logged.