d3dx.ini: Add and document DX10VSCBNumber & DX10DSCBNumber
My experiments have paid off and I have confirmed that these two options
define which constant buffer the driver uses to pass
-separation*convergence and separation to the vertex and domain shaders
it modifies (the equivalent option for DX9 games is GameSpecific0).

They are needed in any game that already uses cb12 in the vertex/domain
shaders, which would have the symptom of having 2D geometry, but until
now we have depended on NVIDIA to fill these in for us (or in the case
of ARK, using arbitrary resource copying to reassign the constant buffer
on one specific affected shader).