Custom resources: Add a mode option to force stereo or mono
The arbitrary resource copying can force the surface creation mode, but
only when a resource is created as a result of copying another, not when
substantiating one from scratch. This can be worked around by
substantiating one resource then copying it to another with the desired
keyword, but that is not very intuitive.

This adds a mode option to the Resource sections to specify the surface
creation mode to use when creating the resource, and as an alternate to
specifying it during the copy (this does not deprecate specifying it
during the copy as that is still the way to do it when copying directly
to a pipeline slot, and the only time that the stereo2mono mode makes
sense).