Custom resources: Add bind flag override
It seems that the bind flags are used by the driver heuristics to
determine whether to stereoise resources or not. e.g. a buffer with a
render target bind slot may be stereoised, and a buffer with a UAV bind
slot will be stereoised if stereo compute shaders are enabled in the
profile.

That means that there might be some circumstances where we might want to
set specific bind flags to influence the driver heuristics to do what we
want, so add the ability to override them.