Custom resources: Allow buffers to be loaded from a file
This extends the filename option of custom resource sections to work
with buffers as well. For these, the buffer type must be specified in
the section as this cannot be determined automatically like it can from
images. The format or structure byte stride should also be set as
required for the type of buffer selected - only the byte width is set
automatically, though this can still be overridden if desired.