Resource copying: Add a region_copy keyword for buffers
In exploring driver heuristics I wondered if the type of copy operation
performed might influence the driver, so added a new region_copy keyword
to force 3DMigoto to use a region copy on buffers even in circumstances
where a full copy would do. This did not seem to make a difference to
the heuristics for my immediate goals, but might still be useful for the
future.