cmd_Decompiler: Reconstruct and validate sections separately
The --validate option was treating the entire shader as one, using
Flugan's trick of copying non-bytecode sections to the new shader. It is
not particularly useful to validate that memcpy works, but it could be
useful to validate that the signature sections are constructed
accurately. Change the validation routine to assemble the shader from
scratch, then compare each section separately, reporting data
mismatches, size mismatches and any missing sections (the later are not
considered fatal).