Assembler: Add missing PS input case found in WATCH_DOGS2
WATCH_DOGS2 has shaders that fail with the assembler. Comparing one such
shader to the original by disassembling, assembling and disassembling
it again with these commands:

    cmd_Decompiler.exe --disassemble-ms 005e4125bef404a6-ps.bin
    cmd_Decompiler.exe -d 005e4125bef404a6-ps.bin
    cmd_Decompiler.exe -a 005e4125bef404a6-ps.asm
    cmd_Decompiler.exe -d 005e4125bef404a6-ps.shdr
    diff -u 005e4125bef404a6-ps.msasm 005e4125bef404a6-ps.asm

Shows that this line has been mangled:

+dcl_input_ps_siv linear noperspective sample v5.xy, position
-dcl_input_ps_siv linear noperspective s0.,

The "sample" keyword was not being handled by the assembler. Use the
improved validation mode of cmd_Decompiler to identify the correct value
to place in the instruction:

    $ cmd_Decompiler.exe -d --validate 005e4125bef404a6-ps.bin

    Disassembling (Flugan) 005e4125bef404a6-ps.bin...
    Parsing Input Signature section...
    Parsing Output Signature section...
    Manufacturing placeholder SHEX section...
    Reassembled shader missing RDEF section

      Section SHEX 0x00000004: expected 0xa1, found 0xa0
      Section SHEX 0x000000fd: expected 0x38, found 0x20
    ...

The mismatch at offset 0x4 is a length, which is one less and to be
expected given the assembler generated one to few instructions. It's not
immediately clear what shift we need to apply, but a quick experiment
setting the entire _11_23 field to 0x1fff reveals which bits correspond
to that field:

      Section SHEX 0x000000fd: expected 0x38, found 0xf8
      Section SHEX 0x000000fe: expected 0x00, found 0xff

The value 0xf8 indicates we need to shift 0x38 by 3 (8+3=11?), which
gives 7. Plug that in to the field and we get a successful verification:

    Disassembling (Flugan) 005e4125bef404a6-ps.bin...
    Parsing Input Signature section...
    Parsing Output Signature section...
    Manufacturing placeholder SHEX section...
    Reassembled shader missing RDEF section
    Reassembled shader missing STAT section
        Assembly verification pass succeeded
      -> 005e4125bef404a6-ps.asm