Assembler: Fix resource type declarations
All the resource declarations share a common opcode (88), are all
documented on the one MSDN article, all have a common means of
representing the format and all support the same formats.

So add a helper to do this in one place, consistently and handling all
formats with confirmed values. This fixes a hang in WATCH_DOGS2 and
Song of the Deep.