Assembler: Note various missing combinations of instructions
The assembler seems to be only designed to cope with the set of
instructions seen in the wild, and not all possible variations of those
instructions. This is starting to bite us, so go through and quickly
audit the code looking for any obviously missing combinations and tag
them with a FIXME. We don't need to wait to find these in a game - they
are all quite trivial to write a testcase in HLSL to generate them.