Assembler: Handle arbitrary MSAA samples
MSAA textures and texture arrays can specify the number of samples in
their declaration, which is encoded into the instruction. Currently the
assembler only handles a small subset of possible values, which is
asking for trouble the next time a game uses a different value.

The encoding is trivial, so parse it properly and encode it properly.