Assembler: Add support for double precision resources and instructions
This adds support for loading from double precision resources (both
double and double2), as well as all the double precision instructions,
most of which have been tested with resource_types.hlsl.

The two global flags relevant to double precision shaders are also
added, and the code to parse these is changed to a loop rather than
processing a fixed number of arguments.