Assembler: More auditing of implemented instructions
Many of the instructions are implemented elsewhere in the assembler from
the instruction table, and not all assembly text matches the names from
the binary decompiler, so it is difficult to see which of these are
implemented and which are missing.

Change all hardcoded opcodes throughout the code to hex so we can match
them up with the table, then go through each instruction missing from
the table and see if it is implemented elsewhere or not, fixing the name
in the comment of the table if it is, and making sure the TODO comments
are on the right lines.

It also looks like the various load instructions have (up to) four
variants each: a basic variant in the instruction table, aoffimmi,
indexable and aoffimmi_indexable variants in the load table. These are
differentiated by a field in the instruction that appears to be quite
consistent, but many of the permutations are missing at the moment. I am
certain these will be safe to add, but just tag them with TODO comments
in this commit in case we need to bisect a breakage later since this
commit is not supposed to make any functional changes.