Assembler: Add support for more permutations of resource loading
Most of the resource load instructions include up to four variations: A
basic version, indexable (used when compiling the shader for shader
model 5), aoffimmi (address offset immediate) and aoffimmi_indexable,
but a number of the permutations of these were missing.

Add all the permutations that have been verified using the
resource_types.hlsl test shader. Also add a couple of cases of the basic
instruction that could not be verified - these seem to only be relevant
to shader model 5, which usually uses the indexable variants, but I
don't know that there is a guarantee that it won't ever use the
non-indexable variants, so add them for safety - they should not hurt
even if they cannot be used.

Remove the TODO comment from load operations that I am convinced cannot
have an aoffimmi variant, and add an exaplanation for why.

Add some commented out code to support the _rcpFloat variant of the
resinfo instruction - I'm fairly certain this is correct, but I haven't
worked out what the appropriate HLSL is to generate it.