Assembler: Add support for dcl_gsinstances
Found a nice trick to find what attributes are valid for a given shader
type - intentionally make one invalid [foo], then check the warning from
fxc:

    warning X3554: unknown attribute foo, or attribute invalid for this
    statement, valid attributes are: maxvertexcount, MaxVertexCount,
    instance

So, add an [instance(n)] attribute to the sv_gsinstanceid.hlsl test
shader, which produces a dcl_gsinstances declaration that the assembler
didn't support. This one is a simple declaration and works fine if
implemented in the instruction table.