ISGN parsing: Invert used mask for system values using SPRs
Validation of WATCH_DOGS2 shaders located an issue where the used mask
was not being correctly inverted in the ISGN sections for system values
that use special purpose registers. This was one case missed while
testing as I could only find these cases in the OSGN section.

Extend the invert case to work with these special values, and invert the
value for 'YES' - since this was only used in outputs until now I had
directly returned the inverted value, but the correct action is to start
with 1 and invert it for outputs.