Assembler: Fix missing viewport array index input to the pixel shader
Looks like all the _siv declarations parse the same system values, but
inconsistently (not all system values are appropriate to the input and
output of each shader, but who cares - this is an assembler, not a
semantic validator). We should probably fix this so they all use common
system value parsing, but this is slightly complicated by the fact the
syntax of these declarations differs and the parsing logic in
dcl_input_ps_siv is pretty... expectant.

For now, just add viewport_array_index to dcl_input_ps_siv since one of
the signature test cases was failing.