Assembler: Change most system value declarations to use common code
All dcl_*_s?v instructions declare system values for use in the shader,
and the system values use the same keywords in all of them, so move that
out into a common routine to catch all possible system values in one
fell swoop.

dcl_input_ps_siv is not completely converted since it's logic is not
straight forward, and dcl_output_sgv is not implemented at all.