Assembler: Add emit_then_cut and emit_then_cut_stream
Neither of these instructions seem to ever be produced by fxc (but that
does not mean they can never be produced with a future compiler, an old
one, a different one, or under certain optimisation scenarios), so we
can't use a HLSL shader to generate them and validate the assembler
against it, so instead just implement them how we expect them to be
encoded and verify that assembling and disassembling the shader works
and produces the same instruction text.

Add the assembly shaders used to test this to the test shaders
directory. Also extend the HLSL shaders such that they should generate
these instructions (but don't).