Assembler: Implement _sat_ugroup variants of the sync instruction
This implements the last four variants of the sync instruction mentioned
on the MSDN page:

https://msdn.microsoft.com/en-us/library/windows/desktop/hh447241(v=vs.85).aspx

I have not managed to get fxc to produce these, which is understandable
since they are lighter weight memory barriers than the global barriers
it seems to prefer, but that does not mean they could not be produced
under the right optimisation conditions.

Add test cases for them in assembly and use that to validate the text
that the disassembler produces, which turns out to have a _sat that the
MSDN article failed to mention.