Assembler: Change sync parsing to parse the instruction as flags
Rather than list all the permutations of the sync instruction, change it
to iterate over the flags listed in the instruction filling out the
_11_23 field of the instruction as it finds them. Realistically this is
likely closer to how the hardware operates as each bit in the flags
field will indicate something it needs to synchronise with, and this
seems to work fine assembling and disassembling a test case that covers
all permutations (including several not mentioned on MSDN).

Notably, combining the _ugroup and _uglobal flags is the one case that
is treated specially by the disassembler and turns into _sat_uglobal. It
is not immediately clear what this signifies, but the fact that the
disassembler special cased it suggests that it has a different meaning
to combining _ugroup and _uglobal (which makes sense given that _ugroup
is a weaker version of _uglobal).