Assembler: Add support for resinfo_rcpfloat variant
I haven't managed to get fxc to produce the _rcpfloat variant of the
resinfo instruction - there is no obvious variant of the GetDimensions()
function that maps to it, and trying to use the reciprocal of the result
does not seem to be optimised to use it, but that does not necessarily
mean that it will never be used.

Add an assembly test case and verify that it assembles and disassembles
back to the same text - note that MSDN refers to it as _rcpFloat, but
the disassembler used all lower case _rcpfloat.