Assembler: Add support for uaddc and usubb instructions
These instructions are (I guess) used to detect overflows and underflows
when performing unsigned adds and subtracts, but there doesn't seem to
be a HLSL intrinsic that maps to them and it is not clear what (if any)
combination of HLSL would be optimised to use them. Therefore, add an
assembly test case and validate that it assembles and disassembles back
to the same text.