Test shaders: Add a test case for the debug layer instructions
Implementing the debug layer instructions is not a high priority since
we can't ship a shader to an end user that uses them, but it might still
be a (very) useful way to get arbitrary information out of a shader
while debugging it - but this will be most useful if we can run with the
debug layer enabled and have the messages go to our log file (e.g. by
using ID3D11InfoQueue to get the messages), so for now just add this
test case as something to work towards.

Move the abort test case from compute.hlsl into this one as that is part
of the debug layer.