Resource copying: Failed attempt to implement stereo2mono for buffers
This was an attempt to support the reverse stereo blit for buffers as
well as 2D textures, but it did not work - it just copied the original
stereo buffer, which remained stereo, and did not contain the 2nd copy
of the buffer.

It is conceivable that this just failed due to a driver bug and may work
in the future, so consign it off to a topic branch.