Fix logging wraparound.
ClearRenderTargetView had missing \n, so it wrapped log, and hid a Present call.

This adds the \n there.

This also adds a space before every \n in the HackerContext, so that it's more clear where we have them, and where we don't. Having them buried at the end of the line has created multiple bugs that needed fixing.

Specifically avoids adding \n to other FrameAnalysisLogs.