Fix crash with Elite Dangerous.
Game actually Releases the HackerContext it got from CreateDevice, which we did not expect.  That passed null as the HackerContext to the Overlay, and later to RunFrameActions causing the crash.

This fix is pretty much game specific, but won't cause any problems when games act normal.

Checks for a null context being passed in when creating the SwapChain, and if so, does pDevice->GetImmediateContext in order to generate a new HackerContext.

The game also calls GetImmediateContext right after creating the SwapChain, so instead of creating it then, it will now receive the reference created in the HackerDXGISwapChain constructor.