Update IDXGIDevice1 handling for Batman: telltale.
It makes no sense to make another object, the stupid QueryInterface is just a low-budget casting mechanism. Rather than make a different object, this upconverts any use of IDXGIDevice to IDXGDevice1 and returns that whenever either is requested.

These are both part of Win7 base platform, as described in the new object table at the top.  So, there is no reason to use anything except for IDXGIDevice1.

Updates the ID3D11Device1 creation to match the approach used by other code.