Add platform update pieces.
Starting with a table of objects at the top, detailing what is and what is not included with platform update. The whole object1 naming is completely lame and confusing, so this details every object level and when to expect them.

Unifies the code for GetParent to use our standard code approach, which removes the need for nullptr checks by having an early Failed check.

Add some basic object handlers for platform update objects that are stubs for now.