On platform update, create HackerDevice1 and HackerContext1.
If enabled, we'll make those objects in CreateDevice, and CreateDeviceAndSwapChain.  They are stored in the same HackerDevice and HackerContext references, as a compatible object.

Any subset calls will still work, and any superset calls will work if the object is coerced using QueryInterface.

In QueryInterface if we see those objects, we return 'this'.