Input: Allow key bindings to be inverted
This allows key bindings to be prefixed with no_ to activate the key
binding only when they are not pressed. This allows us to add bindings
that avoid conflicts with driver bindings, e.g. "no_alt f1" allows for a
key binding on F1 that will not activate when taking a 3D screenshot.

"no_modifiers" is added as a shorthand for "no_ctrl no_alt no_shift
no_lwin no_rwin".