d3dx.ini: Add no_modifiers to all default key bindings
This avoids conflicts between F11 to cycle SBS output modes,
Ctrl+Alt+F11 to toggle compatibility mode and Ctrl+F11 to cycle the
frustum adjustment.

The other bindings are probably less important to exclude modifiers
since most don't conflict with the default driver bindings (except F7
that could conflict with save in-game changes on Ctrl+F7), but Tonka did
request this to avoid conflicts with his custom bindings, and in general
it seems like the right thing to do.