Command Lists: Split CommandList and CustomShader parsing into two stages
Since these two sections can be referred to from any other command list
section, including other sections of the same type, the list of
available sections should be constructed before parsing the contents so
that the section data structures will be guaranteed to exist by the time
we parse the run commands that refer to them - currently there is no
such guarantee and referring to one CommandList from another will only
work if the second has already been parsed (which would require it to
come earlier alphabetically).

Split the parsing of these sections in two - in the first pass just
enumerate the available sections and construct the data structures in
the global maps, and in the second pass actually parse their contents to
fill them in. This allows these sections to refer to each other without
any limitations.