Command Lists: Add a recursion limit
Now that command lists are parsed in two separate phases, they can
potentially refer to themselves, or a cycle could be created leading to
infinite recursion until the stack runs out and the game crashes.

Add a recursion limit to prevent this from occurring due to a simple
typo. The recursion limit is set to 64, which should be more than
generous enough for anyone's legitimate needs... unless they are doing
something truly crazy ;-)