Special heuristic for WATCH_DOGS2 1.09.154 update
Since WATCH_DOGS2 has a pixel density option, the swap chain resolution
is useless so get the resolution from the depth buffer instead. As of
the 1.09.154 update, the game has started creating ridiculously sized
rectangular shadow maps (16384 x 4096 on ultra) which get mistaken for
the resolution.

Refactor the heuristics to guess if a depth buffer might match the
resolution into their own function for clarity, and add a special case
to exclude the aspect ratio of exactly 4, which I don't think would ever
be used as a valid resolution. We could potentially use an aspect ratio
cap, but we would have to be careful of 3D Vision Surround resolutions
that are greater than that.