Experimental command line override support
This is an experiment to see if we can override the command line from
3DMigoto to try to simplify the installation instructions for games that
require it. It hooks into the GetCommandLineA and GetCommandLineW
functions and allows us to append a value to them.

Unfortunately, while it technically does work in most cases d3d11.dll
will be initialised too late for the game to get the overridden value.
The code may still be useful for inclusion into a helper library that is
loaded sooner (preferably in a DLL that the game is directly linked
against rather than one loaded programatically to ensure that the code
runs early enough).

It is not clear if this approach would work at all for games that use
argc/argv from their WinMain functions.