In CreateFactory1, actually create Factory2, if platform_update.
If platform_update, we want to return the highest level object supported, which is IDXGIFactory2, and properly wrapped as HackerDXGIFactory2.

This adds that sequence to CreateFactory1, to possibly create the Factory2 when specified.

Reorders the Create functions, so that the higher level real factory calls are available.

Untested, but compiles.