Fix returning DXGIFactory2 for platform update.
Had the wrong idea there, Factory2 does not have an entry point in the dxgi.dll.  You can only create one by passing the right riid=IDXGIFactory2 to either CreateFactory or CreateFactory1.

So for both those case, this now looks for platform update and builds a Factory2 if it's enabled, always.  That's a superset of Factory1, and should still work, even if they only wanted Factory1.

The WrapFactory2 function will use the sure to be there call of fnOrigCreateFactory1.

Tested in Dishonored, and it successfully creates Factory2, and following calls work.