Return 'this' for HackerDXGIFactory::QueryInterface.
Previously we were wrapping the result from QueryInterface, but that makes extra copies of the same object that is likely already there.

This change determines if the current object is already a HackerDXGIFactory2 using dynamic_cast, and just returns 'this' for that case.

This is part of the rework to return objects we have already created, rather than extra wrapped objects.