Add IDXGIAdapter1, IDXGIDevice2, and IDXGIResource.
For platform update support, it looks like we need the IDXGIAdapter1 and IDXGIDevice2 objects.  Adding those as wrapped objects here, so that we do not return an error upon those requests.

Also adding IDXGIResource and Resource1 as logged objects.  Not sure this is at all necessary, but should not hurt.