Back out last changes, so that input params are modified.
There doesn't seem to be a good way to manage the ViewPorts or Stencil in 1x width, and it makes more sense to have the caller choose a good drawing rect instead for each eye.

This restores the force 2x for DepthStencil, ViewPort, to match the 2x backbuffer.

This matches DarkStarSword's approach for the UE plugin as well.