Add \n for logging problems.
Not sure if these should be no \n for FrameAnalysis purposes, but these Begin/End calls continue to wrap badly in the logs when I'm analyzing problems, and I just need this.

Apologies in advance if this make FrameAnalysisLog look worse.

Longer term goal will be to make the 99% case work with no \n, with a different LogInfo/LogDebug.  That refactoring should be done in a new 1.3 branch though, because it wll mostly break old code comparisons with too much noise.