Once more, with completeness.  Fix all missing \n.
OK, got a grip here, and did a search for all missing \n in HackerContext, the only file I'm willing to currently add the \n to Log calls.

Did a global search for all missing entries, and added them, so log will be consistent.

Also restored C standard \n, with no leading space, as a current compromise.  Not sure this is good, it adds a lot of line noise for future comparisons, but until we get around to refactoring LogInfo and LogDebug, this is probably better?

Really not sure, it adds a lot of line changes, which will make future diffs messier.  If this seems wrong, or annoying, please revert.

Big problem was actually missing \n on a lot of calls, leading to my confusion.