Add the D3D11_BIND_RENDER_TARGET check for CreateTexture2D.
We need the rendertargetview to match the depthstencil for a given OMSet, and were not doing the rendertargetviews at all.

This adds the check in CreateTexture2D for D3D11_BIND_RENDER_TARGET as well, and will modify the width for full screen creation in either case.

Works correctly in the Unity deferred rendering test case.