Rename MigotoException to 3DmigotoException
We try not to use acronyms and contractions because they reduce clarity and make it hard to read.  The project 'proper' name is 3Dmigoto, with a lower case 'm'.  Chiri's choice, just trying to respect his original contribution.

Also removed the crc-32 full path.  We specifically do not want those clumsy paths as part of our include files.  If it doesn't compile, that means the project is misconfigured, and needs to be fixed in the VS project file paths instead.  

Also added trailing CR for the file, as GitHub does not like files that end without CR.