Change most uses of GetPrivateProfileInt to use GetIniInt
We are going to switch out the underlying ini parsing API, so we want
everything to go through our helper routines to make this easier. Switch
remaining uses of GetPrivateProfileInt to GetIniInt, except for those
that have a comment indicating that they should not use the helper
because logging has not yet been enabled, which is currently required by
the helper function. We can worry about those exceptions later if need
be, but since they are only one call per ini load (two if the profile
manager helper is involved) their performance impact is likely minimal
so they are of little concern.