IniHandler: Add our own ini parser and switch GetIniString to use it
This opens the d3dx.ini and reads it into our own data structures, and
switches calls to GetIniString over to use it. The section parser still
uses the old APIs, but this already cuts the config reload time in
Dreamfall Chapters from 1:45 down to just 15 seconds.

It tries to replicate a lot of the same behaviour as
GetPrivateProfileString just in case we have any fixes that rely on
any quirks of that function. In particular:
- Section headers start with a [, but the terminating ] is optional
- Only the first ] is used in a section header, everything after that is
  ignored
- Only the first section with a given (case insensitive) name is used
  when looking up keys (the other APIs do parse subsequent sections)
- If a key is duplicated within a section, the first occurrence is
  returned when looking up the key by name (other APIs support duplicate
  keys)
- A line like "foo=" has an empty string for the value
- A line like "foo" will not be found with GetPrivateProfileString

We explicitly don't care about some edge cases, like the completely
different implementations of the API if either section or key are passed
as NULL (which change that API to behave like some of the other APIs in
the same family that do completely different things), or keys that
appear outside of a section.

An optional validator can be enabled by defining VALIDATE_INI_PARSER.
This will call both old and new APIs and compare their results to find
these type of quirks. Obviously, this validator should be disabled for
production since it will incur the costs of both APIs.

We continue using the old API to check for [Logging] calls=1, so that we
can have the log file open during our parser to warn about duplicate
sections. Since that is only a single call it should have a minimal
performance impact (and I'm pretty sure the old API stops parsing the
moment it finds the entry it is looking for, and since our templates
have logging at the start of the d3dx.ini, this should be very minimal).