Ini Handler: Implement equivalent of GetIniSections in ParseIni
Fill out a sorted version of the sections while parsing the ini file,
which removes the need to call the old GetPrivateProfileSectionNames API
and deal with it's quirks. This also moves the duplicate section warning
into the new parsing routine.

For now this still calls GetIniSection since that adds a bunch of
validation, but we will need to separate that validation out to remove
that dependency.

Like last time, we add a validation routine activated by defining
VALIDATE_INI_PARSER to validate that the new API produces the same
result as the old one.