Ini Handler: Remove validation and passing iniFile everywhere
The new ini parser seems to be working ok, so remove the validation
routines (if there's anything suspect, you can always revert this or git
checkout the commit before).

Removing the validation routines means that the only two remaining calls
to the old GetPrivateProfile APIs are to check if logging is enabled.
This also means that we no longer need to pass the ini filename around
all over the place, so rip it out of all the function calls that
included it.