Resource copying: Log ini line when [re]creating resource caches
When creating or recreating a cached resource we currently only print a
generic message indicating what we are doing, but we don't print out
anything to identify which resource or copy operation it relates to.

For simple cases, or where the shaderhacker only modifies one thing at a
time it is easy enough to tell that the line they just added was the
cause, but as mods get more complicated or multiple resource copies are
added at once it becomes difficult, and in some cases a resource may
legitimately need to be recreated periodically making it impossible to
spot if any new messages have appeared.

This passes the section name and line from the ini file that the copy
directive relates to into the RecreateCompatible* routines, and changes
the log message to use that to make them identifiable. The new text is
placed near the end of the line to make these logs easily
distinguishable from the ini parsing logs at a glance.

The alternative was to log the resource name, but that only makes sense
when recreating a custom resource, and not when copying a resource to
the pipeline or creating a stereo2mono intermediate resource cache.
Using the line from the ini file should make all of these cases
identifiable without over-complicating the code. The intermediate
stereo2mono resource cache will be tagged with "(intermediate)" to
distinguish them from the destination resource cache.

This also changes the frame analysis log messages of these lines to
include the ini section so that we don't have to store this separately
and can reduce the redundant information we store.

e.g. the new log messages in a Unity game with my template look like:

    Substantiated custom Buffer resource
    Creating cached buffer [Present] resource_hud_depth_cb = copy resource_hud_depth_uav
    Creating cached buffer [ShaderOverride_83bd33475a729417] resource_unityperdraw = vs-cb3
    Substantiated custom Buffer resource
    Creating cached buffer [ShaderOverride_83bd33475a729417] vs-cb10 = copy resource_inverse_mvp
    Creating cached buffer [ShaderOverride_352d8d3c52aa9229] ps-cb10 = copy resource_inverse_mvp
    Creating cached resource [ShaderOverride_93e3a6cda25c688d] resource_cameradepthtexture = ps-t0
    Creating cached buffer [ShaderOverride_93e3a6cda25c688d] resource_unitypercamera = ps-cb1
    Creating cached resource [customshader3dvision2sbs] ps-t100 = stereo2mono bb
    Creating cached resource [customshader3dvision2sbs] ps-t100 = stereo2mono bb (intermediate)